% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metrics_info.R
\name{get_metrics_info}
\alias{get_metrics_info}
\title{Get information on metrics}
\usage{
get_metrics_info()
}
\value{
Summary table of metrics details
}
\description{
Get information on metrics
}
\examples{
get_metrics_info()

}
