% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_xray.R
\name{make_xray}
\alias{make_xray}
\title{Make data xray by a grouping variable}
\usage{
make_xray(data, by = NULL)
}
\arguments{
\item{data}{A data frame.}

\item{by}{Optional name of grouping ("by") variable as character string.}
}
\value{
A tibble containing variable metadata with 1 row per group.
}
\description{
Create comprehensive tibble of variable metadata using Hmisc::describe as engine, with option for grouping
}
\examples{

make_xray(safetyData::adam_adsl)

make_xray(safetyData::adam_adsl, by = "ARM")

}
