% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_xray_core.R
\name{make_xray_core}
\alias{make_xray_core}
\title{Make data xray - core function}
\usage{
make_xray_core(data)
}
\arguments{
\item{data}{A data frame.}
}
\value{
A tibble containing variable metadata.
}
\description{
Create comprehensive tibble of variable metadata using Hmisc::describe as engine.
}
\examples{

diamonds <- ggplot2::diamonds
make_xray_core(diamonds)

}
