% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_nested_tabs.R
\name{nested_tab_counts}
\alias{nested_tab_counts}
\title{Create nested table display for counts}
\usage{
nested_tab_counts(counts_df)
}
\arguments{
\item{counts_df}{Dataframe of COUNTS from Hmisc::describe}
}
\description{
Create nested table display for counts
}
\keyword{internal}
