% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_nested_tabs.R
\name{nested_tab_extremes}
\alias{nested_tab_extremes}
\title{Create nested table display for extremes}
\usage{
nested_tab_extremes(extremes_df)
}
\arguments{
\item{extremes_df}{Dataframe of EXTREMES from Hmisc::describe}
}
\description{
Create nested table display for extremes
}
\keyword{internal}
