% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_nested_tabs.R
\name{nested_tab_theme}
\alias{nested_tab_theme}
\title{Theme for nested tables}
\usage{
nested_tab_theme()
}
\value{
Object of class reactableTheme
}
\description{
Theme for nested tables
}
\keyword{internal}
