% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_nested_tabs.R
\name{nested_tab_values}
\alias{nested_tab_values}
\title{Create nested table display for values}
\usage{
nested_tab_values(values_df)
}
\arguments{
\item{values_df}{Dataframe of VALUES from Hmisc::describe}
}
\description{
Create nested table display for values
}
\keyword{internal}
