% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_xray.R
\name{report_xray}
\alias{report_xray}
\title{Create X-ray report}
\usage{
report_xray(data, by = NULL, data_name, study, loc = NULL)
}
\arguments{
\item{data}{A data frame.}

\item{by}{Optional name of grouping ("by") variable as character string.}

\item{data_name}{Name of dataset to be displayed in report as character string.}

\item{study}{Name of study to be displayed in report as character string.}

\item{loc}{Directory to save the rmd and html output. Defaults to current working directory.}
}
\value{

}
\description{
Create X-ray report
}
\examples{

\dontrun{
 report_xray(safetyData::adam_adsl, by = "ARM", data_name = "ADSL", study = "CDISC_pilot", loc = getwd())
}

}
